requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-dotenv>=1.0.0
numpy>=1.24.0

